fix(hexfield+hexo_models): correctness bundle from full-stack review — search signs, SH integrity, target features, durability
Six-subsystem review of the live main_7 stack; every fix independently
re-verified from source before implementation. All behavior changes are
correctness-restoring; fresh-root/kp=1.0 paths are bit-identical.

Search (rust, hexfield + both hexo_models twins):
- advance_root seeded the promoted root with -edge.value_sum
  unconditionally; on same-player promotions (FirstStone->SecondStone
  keeps current_player, ~half of all moves) the edge perspective already
  matches, so the seed was sign-flipped. Corrupted the exported
  root_value -> short_term_value training targets since epoch 1, plus
  fast-search root FPU. Now negates only on a player change.
- Sequential-halving round advancement ranked survivors on virtual-loss
  contaminated visits/completedQ (barrier counted in-flight selections).
  Now gated on a drained slot (continuous) / drained search (lockstep).
- SH ranking and the exported gumbel target computed the sigma scale
  from cumulative, reuse-inflated max_n; now from this move's delta
  visits over a move-entry snapshot (GumbelRootState.entry_visits).

Targets (python):
- Fast (PCR value-only) rows were written with first_stone/hot/win facts
  force-empty ("fast rows are not written" comment was stale) — ~2/3 of
  value-head rows trained with tactical planes zeroed vs serve. Facts
  now recomputed in the writer thread, bit-identical to the full branch.
- binned_value_loss binned continuous scalar targets in fp16 under
  autocast (~3% of a bin mis-split); targets now fp32 end-to-end.
- rust expand backend validated only the stvalue block WIDTH; now exact
  horizon-tuple equality (kernel copies positionally, cannot remap).

Durability (python; host takes real power cuts):
- save_checkpoint and write_compact_shard now tmp+fsync+os.replace
  (tmp names stay off the supervisor/manifest globs; npz-then-sidecar
  commit order preserved).
- build_window_split skips torn shards loudly instead of crash-looping
  every epoch; selfplay resume counts only npz+sidecar pairs as done
  (keys still advance past uncommitted npz).

Training volume:
- keep_prob was applied on top of the already-row-capped file selection
  (silent ~kp x 96k rows/epoch once the window exceeds 600k rows,
  ~ep120); selection request now inflated by 1/kp and effective_rows
  debits min(requested, selected*kp).

Tests: 4 new rust unit tests (47 pass), 11 new pytest cases; full suite
83 passed / 0 failed (2 pre-existing test_hexfield_model failures
confirmed present at HEAD, tracked separately).

+++ packages/hexfield/python/hexfield/losses.py
@@ -163,11 +163,15 @@ def binned_value_loss(
     Masked rows contribute 0. The denominator defaults to the masked row count
     (or the item count when unmasked) and is overridable."""
 
-    target_tensor = torch.as_tensor(target, device=logits.device, dtype=logits.dtype)
+    # Targets stay fp32 end-to-end: under train autocast ``logits.dtype`` is fp16,
+    # which would quantize continuous scalar targets on entry AND compute the
+    # two-hot position (in [0, 64], fp16 ulp ~1/32 near the top) in fp16,
+    # mis-splitting adjacent-bin weights by up to ~3% of a bin (short-term
+    # values, cell_q, moves_left). The CE below already lifts logits to >= fp32
+    # via ``_at_least_fp32``, so an fp32 target costs nothing.
+    target_tensor = torch.as_tensor(target, device=logits.device, dtype=torch.float32)
     if target_tensor.shape != logits.shape:
-        target_tensor = scalar_to_binned_target(target_tensor).to(
-            device=logits.device, dtype=logits.dtype
-        )
+        target_tensor = scalar_to_binned_target(target_tensor).to(device=logits.device)
     if logits.shape != target_tensor.shape:
         raise ValueError(
             f"value target shape {tuple(target_tensor.shape)} != logits {tuple(logits.shape)}"
